@@ -1,15 +1,18 @@
 [build-system]
-requires = ["maturin>=1.2"]
+requires = ["maturin>=1,<2"]
 build-backend = "maturin"
 
 [project]
 name = "forzium_engine"
-version = "0.1.1"
-description = "ForziumAPI-compatible framework with a Rust core."
+version = "0.1.4"
+description = "FastAPI-compatible framework with a Rust core."
 authors = [{ name = "pavanepour-k" }]
-license = { text = "MIT" }
-requires-python = ">=3.11"
-dependencies = []
+requires-python = ">=3.12,<3.13"
+dependencies = [
+    "graphql-core>=3.2,<4",
+    "pydantic>=2.9,<2.10",
+    "PyYAML>=6,<7",
+]
 
 [tool.maturin]
 bindings = "pyo3"
@@ -19,34 +22,41 @@ python-source = "."
 [tool.pytest.ini_options]
 addopts = ""
 testpaths = ["tests"]
-filterwarnings = [
-    "ignore:cannot collect test class 'TestClient' because it has a __init__:pytest.PytestCollectionWarning",
-    "ignore:'cgi' is deprecated and slated for removal in Python 3.13:DeprecationWarning",
-]
 
 [tool.poetry]
 name = "forzium-engine"
-version = "0.1.1"
-description = "ForziumAPI-compatible framework with a Rust core."
+version = "0.1.4"
+description = "FastAPI-compatible framework with a Rust core."
 authors = ["pavanepour-k"]
 package-mode = false
 
 [tool.poetry.dependencies]
-python = ">=3.11"
+python = ">=3.12,<3.13"
+graphql-core = "^3.2"
+pydantic = ">=2.9,<2.10"
+PyYAML = "^6.0"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^8.4.2"
-pytest-cov = "^6.2.1"
-black = "^25.1.0"
-ruff = "^0.12.11"
-mypy = "^1.17.1"
 bandit = "^1.8.6"
-pip-audit = "^2.9.0"
-safety = "^3.6.1"
-mutmut = "^3.3.1"
-pytest-benchmark = "^5.1.0"
-pip-licenses = "^5.0.0"
+black = "^25.1.0"
 detect-secrets = "^1.5.0"
+grpcio = "^1.60.0"
+grpcio-health-checking = "^1.60.0"
+grpcio-tools = "^1.60.0"
+hypothesis = "^6.0.0"
+mypy = "^1.17.1"
+mutmut = "^3.3.1"
+pip-audit = "^2.9.0"
+pip-licenses = "^5.0.0"
+protobuf = "^4.25.0"
+pytest = "^8.4.2"
+pytest-benchmark = "^5.1.0"
+pytest-cov = "^6.3.0"
+ruff = "^0.12.12"
+safety = "^3.6.1"
 
 [tool.mutmut]
 paths_to_mutate = ["forzium"]
+
+[tool.isort]
+profile = "black"
