@@ -1,4 +1,6 @@
+use crate::compute::rayon_metrics;
 use crate::error::ForziumError;
+use rayon::prelude::*;
 #[cfg(target_arch = "x86_64")]
 use std::arch::x86_64::*;
 
@@ -24,15 +26,21 @@ fn validate_same_shape(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<(usize, usize),
 
 pub fn multiply(m: &[Vec<f64>], factor: f64) -> Result<Vec<Vec<f64>>, ForziumError> {
     validate_matrix(m)?;
-    Ok(m.iter()
-        .map(|r| r.iter().map(|v| v * factor).collect())
+    Ok(m.par_iter()
+        .map(|r| {
+            let _guard = rayon_metrics::track_task();
+            r.par_iter().map(|v| v * factor).collect()
+        })
         .collect())
 }
 
 pub fn add(m: &[Vec<f64>], addend: f64) -> Result<Vec<Vec<f64>>, ForziumError> {
     validate_matrix(m)?;
-    Ok(m.iter()
-        .map(|r| r.iter().map(|v| v + addend).collect())
+    Ok(m.par_iter()
+        .map(|r| {
+            let _guard = rayon_metrics::track_task();
+            r.par_iter().map(|v| v + addend).collect()
+        })
         .collect())
 }
 
@@ -48,79 +56,91 @@ pub fn transpose(m: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
 }
 
 pub fn matmul(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
-    let (rows_a, cols_a) = validate_matrix(a)?;
+    let (_rows_a, cols_a) = validate_matrix(a)?;
     let (rows_b, cols_b) = validate_matrix(b)?;
     if cols_a != rows_b {
         return Err(ForziumError::Validation("shape mismatch".into()));
     }
-    let mut out = vec![vec![0.0; cols_b]; rows_a];
-    for (row_a, out_row) in a.iter().zip(out.iter_mut()) {
-        for (k, val_a) in row_a.iter().enumerate() {
-            let row_b = &b[k];
-            for (out_cell, val_b) in out_row.iter_mut().zip(row_b.iter()) {
-                *out_cell += val_a * val_b;
+    let out: Vec<Vec<f64>> = a
+        .par_iter()
+        .map(|row_a| {
+            let _guard = rayon_metrics::track_task();
+            let mut out_row = vec![0.0; cols_b];
+            for (k, val_a) in row_a.iter().enumerate() {
+                let row_b = &b[k];
+                for (j, val_b) in row_b.iter().enumerate() {
+                    out_row[j] += val_a * val_b;
+                }
             }
-        }
-    }
+            out_row
+        })
+        .collect();
     Ok(out)
 }
 
 pub fn simd_matmul(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
-    let (rows_a, cols_a) = validate_matrix(a)?;
+    let (_rows_a, cols_a) = validate_matrix(a)?;
     let (rows_b, cols_b) = validate_matrix(b)?;
     if cols_a != rows_b {
         return Err(ForziumError::Validation("shape mismatch".into()));
     }
     let bt = transpose(b)?;
-    let mut out = vec![vec![0.0; cols_b]; rows_a];
-    #[cfg(target_arch = "x86_64")]
-    unsafe {
-        for (row_a, out_row) in a.iter().zip(out.iter_mut()) {
-            for (bt_row, out_cell) in bt.iter().zip(out_row.iter_mut()) {
-                let mut k = 0;
-                let mut vsum = _mm_set1_pd(0.0);
-                while k + 2 <= cols_a {
-                    let va = _mm_loadu_pd(row_a.as_ptr().add(k));
-                    let vb = _mm_loadu_pd(bt_row.as_ptr().add(k));
-                    let prod = _mm_mul_pd(va, vb);
-                    vsum = _mm_add_pd(vsum, prod);
-                    k += 2;
+    let out: Vec<Vec<f64>> = a
+        .par_iter()
+        .map(|row_a| {
+            let _guard = rayon_metrics::track_task();
+            let mut out_row = vec![0.0; cols_b];
+            #[cfg(target_arch = "x86_64")]
+            unsafe {
+                for (bt_row, out_cell) in bt.iter().zip(out_row.iter_mut()) {
+                    let mut k = 0;
+                    let mut vsum = _mm_set1_pd(0.0);
+                    while k + 2 <= cols_a {
+                        let va = _mm_loadu_pd(row_a.as_ptr().add(k));
+                        let vb = _mm_loadu_pd(bt_row.as_ptr().add(k));
+                        let prod = _mm_mul_pd(va, vb);
+                        vsum = _mm_add_pd(vsum, prod);
+                        k += 2;
+                    }
+                    let mut buf = [0.0f64; 2];
+                    _mm_storeu_pd(buf.as_mut_ptr(), vsum);
+                    let mut sum = buf[0] + buf[1];
+                    while k < cols_a {
+                        sum += row_a[k] * bt_row[k];
+                        k += 1;
+                    }
+                    *out_cell = sum;
                 }
-                let mut buf = [0.0f64; 2];
-                _mm_storeu_pd(buf.as_mut_ptr(), vsum);
-                let mut sum = buf[0] + buf[1];
-                while k < cols_a {
-                    sum += row_a[k] * bt_row[k];
-                    k += 1;
-                }
-                *out_cell = sum;
             }
-        }
-    }
-    #[cfg(not(target_arch = "x86_64"))]
-    {
-        for (row_a, out_row) in a.iter().zip(out.iter_mut()) {
-            for (bt_row, out_cell) in bt.iter().zip(out_row.iter_mut()) {
-                let mut sum = 0.0;
-                for (val_a, val_b) in row_a.iter().zip(bt_row.iter()) {
-                    sum += val_a * val_b;
+            #[cfg(not(target_arch = "x86_64"))]
+            {
+                for (bt_row, out_cell) in bt.iter().zip(out_row.iter_mut()) {
+                    let mut sum = 0.0;
+                    for (val_a, val_b) in row_a.iter().zip(bt_row.iter()) {
+                        sum += val_a * val_b;
+                    }
+                    *out_cell = sum;
                 }
-                *out_cell = sum;
             }
-        }
-    }
+            out_row
+        })
+        .collect();
     Ok(out)
 }
 
 pub fn elementwise_add(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
-    let (rows, cols) = validate_same_shape(a, b)?;
-    let mut out = vec![vec![0.0; cols]; rows];
-    for (row_out, (row_a, row_b)) in out.iter_mut().zip(a.iter().zip(b.iter())) {
-        for (val_out, (val_a, val_b)) in row_out.iter_mut().zip(row_a.iter().zip(row_b.iter())) {
-            *val_out = val_a + val_b;
-        }
-    }
-    Ok(out)
+    validate_same_shape(a, b)?;
+    Ok(a.par_iter()
+        .zip(b.par_iter())
+        .map(|(row_a, row_b)| {
+            let _guard = rayon_metrics::track_task();
+            row_a
+                .par_iter()
+                .zip(row_b.par_iter())
+                .map(|(val_a, val_b)| val_a + val_b)
+                .collect()
+        })
+        .collect())
 }
 
 pub fn simd_elementwise_add(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
@@ -156,16 +176,21 @@ pub fn simd_elementwise_add(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<Vec<Vec<f6
 }
 
 pub fn hadamard(a: &[Vec<f64>], b: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
-    let (rows, cols) = validate_same_shape(a, b)?;
-    let mut out = vec![vec![0.0; cols]; rows];
-    for (row_out, (row_a, row_b)) in out.iter_mut().zip(a.iter().zip(b.iter())) {
-        for (val_out, (val_a, val_b)) in row_out.iter_mut().zip(row_a.iter().zip(row_b.iter())) {
-            *val_out = val_a * val_b;
-        }
-    }
-    Ok(out)
+    validate_same_shape(a, b)?;
+    Ok(a.par_iter()
+        .zip(b.par_iter())
+        .map(|(row_a, row_b)| {
+            let _guard = rayon_metrics::track_task();
+            row_a
+                .par_iter()
+                .zip(row_b.par_iter())
+                .map(|(val_a, val_b)| val_a * val_b)
+                .collect()
+        })
+        .collect())
 }
 
+#[allow(clippy::needless_range_loop)]
 pub fn conv2d(input: &[Vec<f64>], kernel: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, ForziumError> {
     let (rows, cols) = validate_matrix(input)?;
     let (krows, kcols) = validate_matrix(kernel)?;
@@ -175,11 +200,12 @@ pub fn conv2d(input: &[Vec<f64>], kernel: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, 
     let out_rows = rows - krows + 1;
     let out_cols = cols - kcols + 1;
     let mut out = vec![vec![0.0; out_cols]; out_rows];
-    #[cfg(target_arch = "x86_64")]
-    unsafe {
-        for r in 0..out_rows {
-            for c in 0..out_cols {
-                let mut sum = 0.0;
+    out.par_iter_mut().enumerate().for_each(|(r, out_row)| {
+        let _guard = rayon_metrics::track_task();
+        for c in 0..out_cols {
+            let mut sum = 0.0;
+            #[cfg(target_arch = "x86_64")]
+            unsafe {
                 for kr in 0..krows {
                     let mut kc = 0;
                     let mut vsum = _mm_set1_pd(0.0);
@@ -198,27 +224,22 @@ pub fn conv2d(input: &[Vec<f64>], kernel: &[Vec<f64>]) -> Result<Vec<Vec<f64>>, 
                         kc += 1;
                     }
                 }
-                out[r][c] = sum;
             }
-        }
-    }
-    #[cfg(not(target_arch = "x86_64"))]
-    {
-        for r in 0..out_rows {
-            for c in 0..out_cols {
-                let mut sum = 0.0;
+            #[cfg(not(target_arch = "x86_64"))]
+            {
                 for kr in 0..krows {
                     for kc in 0..kcols {
                         sum += input[r + kr][c + kc] * kernel[kr][kc];
                     }
                 }
-                out[r][c] = sum;
             }
+            out_row[c] = sum;
         }
-    }
+    });
     Ok(out)
 }
 
+#[allow(clippy::needless_range_loop)]
 pub fn max_pool2d(input: &[Vec<f64>], size: usize) -> Result<Vec<Vec<f64>>, ForziumError> {
     let (rows, cols) = validate_matrix(input)?;
     if size == 0 || rows % size != 0 || cols % size != 0 {
@@ -227,7 +248,8 @@ pub fn max_pool2d(input: &[Vec<f64>], size: usize) -> Result<Vec<Vec<f64>>, Forz
     let out_rows = rows / size;
     let out_cols = cols / size;
     let mut out = vec![vec![0.0; out_cols]; out_rows];
-    for r in 0..out_rows {
+    out.par_iter_mut().enumerate().for_each(|(r, out_row)| {
+        let _guard = rayon_metrics::track_task();
         for c in 0..out_cols {
             let mut m = f64::NEG_INFINITY;
             for pr in 0..size {
@@ -238,9 +260,9 @@ pub fn max_pool2d(input: &[Vec<f64>], size: usize) -> Result<Vec<Vec<f64>>, Forz
                     }
                 }
             }
-            out[r][c] = m;
+            out_row[c] = m;
         }
-    }
+    });
     Ok(out)
 }
 
@@ -277,5 +299,31 @@ mod tests {
         let b = vec![vec![1.0], vec![2.0], vec![3.0]];
         let err = matmul(&a, &b).unwrap_err();
         matches!(err, ForziumError::Validation(_));
+    }
+
+    #[test]
+    fn matmul_parallel_speedup() {
+        use rayon::ThreadPoolBuilder;
+        use std::time::Instant;
+
+        let size = 150;
+        let a = vec![vec![1.0; size]; size];
+        let b = vec![vec![1.0; size]; size];
+
+        let pool1 = ThreadPoolBuilder::new().num_threads(1).build().unwrap();
+        let single = pool1.install(|| {
+            let start = Instant::now();
+            matmul(&a, &b).unwrap();
+            start.elapsed()
+        });
+
+        let pool4 = ThreadPoolBuilder::new().num_threads(4).build().unwrap();
+        let parallel = pool4.install(|| {
+            let start = Instant::now();
+            matmul(&a, &b).unwrap();
+            start.elapsed()
+        });
+
+        assert!(parallel < single);
     }
 }
