@@ -1,14 +1,31 @@
+# flake8: noqa
 """Minimal HTTP primitives with header and cookie support."""
 
 from __future__ import annotations
 
 import asyncio
-import cgi
 import inspect
 import json
 from http.cookies import SimpleCookie
 from typing import Any, AsyncIterator, Callable, Dict, List, Mapping, Tuple, cast
+from types import SimpleNamespace
 from urllib.parse import parse_qs, urlsplit
+
+
+def _parse_header(line: str) -> tuple[str, Dict[str, str]]:
+    """Parse a Content-Disposition header.
+
+    This replaces :func:`cgi.parse_header` removed from the standard library.
+    Only the subset required for multipart form parsing is implemented.
+    """
+    parts = [p.strip() for p in line.split(";") if p]
+    value = parts[0] if parts else ""
+    params: Dict[str, str] = {}
+    for item in parts[1:]:
+        if "=" in item:
+            k, v = item.split("=", 1)
+            params[k.strip()] = v.strip().strip('"')
+    return value, params
 
 
 class Request:
@@ -21,6 +38,8 @@ class Request:
         body: bytes = b"",
         headers: Mapping[str, str] | None = None,
         path_params: Mapping[str, str] | None = None,
+        max_upload_size: int | None = None,
+        allowed_mime_types: set[str] | None = None,
     ) -> None:
         self.method = method.upper()
         self.url = url
@@ -36,6 +55,9 @@ class Request:
         self._form_data: dict[str, Any] | None = None
         self._files: dict[str, dict[str, Any]] | None = None
         self._stream_consumed = False
+        self.max_upload_size = max_upload_size
+        self.allowed_mime_types = allowed_mime_types
+        self.state: SimpleNamespace = SimpleNamespace()
 
     async def body(self) -> bytes:
         """Return the request body."""
@@ -80,14 +102,34 @@ class Request:
                         ),
                         "",
                     )
-                    _, params = cgi.parse_header(disp)
+                    ctype_header = next(
+                        (
+                            h
+                            for h in headers
+                            if h.lower().startswith("content-type")
+                        ),
+                        "",
+                    )
+                    mime = ctype_header.split(":", 1)[1].strip().lower() if ctype_header else ""
+                    _, params = _parse_header(disp)
                     name = params.get("name", "")
                     filename = params.get("filename")
                     content = content.rstrip(b"\r\n")
                     if filename:
+                        if (
+                            self.allowed_mime_types is not None
+                            and mime not in self.allowed_mime_types
+                        ):
+                            raise ValueError("unsupported media type")
+                        if (
+                            self.max_upload_size is not None
+                            and len(content) > self.max_upload_size
+                        ):
+                            raise ValueError("file too large")
                         files[name] = {
                             "filename": filename,
                             "content": content,
+                            "content_type": mime,
                         }
                     else:
                         form[name] = content.decode()
@@ -127,19 +169,21 @@ class Request:
 class Depends:
     """Wrapper marking a callable as a dependency."""
 
-    def __init__(self, dependency: Callable[..., Any]) -> None:
+    def __init__(self, dependency: Callable[..., Any] | None = None) -> None:
         self.dependency = dependency
 
 
 def solve_dependencies(
     dependencies: List[Tuple[str, Callable[..., Any]]],
     overrides: List[Dict[Callable[..., Any], Callable[..., Any]]],
+    request: Request | None = None,
 ) -> tuple[dict[str, Any], list[Tuple[Callable[[], Any], bool]]]:
     """Resolve *dependencies* applying override mappings.
 
-    Returns a tuple of ``(values, cleanup)`` where ``values`` maps parameter
-    names to resolved values and ``cleanup`` contains callables executed after
-    the request to tear down context managers or generators.
+    ``request`` is passed to dependency callables that expect a ``Request``
+    argument.  Returns a tuple of ``(values, cleanup)`` where ``values`` maps
+    parameter names to resolved values and ``cleanup`` contains callables
+    executed after the request to tear down context managers or generators.
     """
 
     cache: Dict[Callable[..., Any], Any] = {}
@@ -155,6 +199,12 @@ def solve_dependencies(
         sig = inspect.signature(actual)
         kwargs: dict[str, Any] = {}
         for param in sig.parameters.values():
+            if param.name == "request" or (
+                param.annotation is Request
+                or (isinstance(param.annotation, str) and param.annotation == "Request")
+            ):
+                kwargs[param.name] = request
+                continue
             default = param.default
             if isinstance(default, Depends):
                 kwargs[param.name] = resolve(default.dependency)
@@ -199,16 +249,11 @@ def solve_dependencies(
 class BackgroundTask:
     """Run *func* after the response is sent."""
 
-    def __init__(
-        self,
-        func: Callable[..., Any],
-        *args: Any,
-        **kwargs: Any,
-    ) -> None:
+    def __init__(self, func: Callable[..., Any], *args: Any, **kwargs: Any) -> None:
         self.func = func
         self.args = args
         self.kwargs = kwargs
-        self.is_async = asyncio.iscoroutinefunction(func)
+        self.is_async = inspect.iscoroutinefunction(func)
 
     async def __call__(self) -> None:
         if self.is_async:
@@ -220,8 +265,13 @@ class BackgroundTask:
 class BackgroundTasks:
     """Collection of background tasks."""
 
-    def __init__(self, tasks: list[BackgroundTask] | None = None) -> None:
+    def __init__(
+        self,
+        tasks: list[BackgroundTask] | None = None,
+        queue: Any | None = None,
+    ) -> None:
         self.tasks = list(tasks) if tasks else []
+        self.queue = queue
 
     def add_task(
         self,
@@ -229,7 +279,10 @@ class BackgroundTasks:
         *args: Any,
         **kwargs: Any,
     ) -> None:
-        self.tasks.append(BackgroundTask(func, *args, **kwargs))
+        if self.queue is not None:
+            self.queue.enqueue(func, *args, **kwargs)
+        else:
+            self.tasks.append(BackgroundTask(func, *args, **kwargs))
 
     async def __call__(self) -> None:
         for task in self.tasks:
