@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import json
+import re
 from dataclasses import dataclass
 from typing import Any, Mapping
 from urllib.parse import urlencode
@@ -17,6 +18,8 @@ class Response:
     status_code: int
     text: str
     headers: Mapping[str, str]
+    content: bytes
+    chunks: list[str] | None = None
 
     def json(self) -> Any:
         """Return the body parsed as JSON."""
@@ -26,6 +29,8 @@ class Response:
 class TestClient:
     """Execute requests against a ``ForziumApp`` without a server."""
 
+    __test__ = False  # prevent Pytest from treating this as a test case
+    
     def __init__(self, app: ForziumApp) -> None:
         self.app = app
 
@@ -37,50 +42,126 @@ class TestClient:
         json_body: Mapping[str, Any] | None = None,
         body: bytes | None = None,
         params: Mapping[str, Any] | None = None,
+        headers: Mapping[str, str] | None = None,
     ) -> Response:
         """Send an HTTP request and return the response."""
-        route = next(
-            (r for r in self.app.routes if r["method"] == method and r["path"] == path),
-            None,
-        )
+        def match_path(
+            template: str, concrete: str
+        ) -> tuple[bool, tuple[str, ...]]:
+            if "{" not in template:
+                return template == concrete, ()
+            pattern_parts: list[str] = []
+            idx = 0
+            length = len(template)
+            while idx < length:
+                if template[idx] == "{":
+                    end = template.find("}", idx)
+                    if end == -1:
+                        pattern_parts.append(re.escape(template[idx:]))
+                        idx = length
+                        break
+                    pattern_parts.append(r"([^/]+)")
+                    idx = end + 1
+                    continue
+                next_brace = template.find("{", idx)
+                if next_brace == -1:
+                    next_brace = length
+                pattern_parts.append(re.escape(template[idx:next_brace]))
+                idx = next_brace
+            pattern = "^" + "".join(pattern_parts) + "$"
+            match = re.match(pattern, concrete)
+            if match is None:
+                return False, ()
+            return True, match.groups()
+
+        route = None
+        path_params: tuple[str, ...] = ()
+        for r in self.app.routes:
+            if r["method"] != method:
+                continue
+            matched, values = match_path(r["path"], path)
+            if matched:
+                route = r
+                path_params = values
+                break
         if route is None:
             raise ValueError(f"no route for {method} {path}")
-        handler = self.app._make_handler(  # pylint: disable=protected-access
+        route_app = route.get("app", self.app)
+        overrides = [route.get("dependency_overrides", {})]
+        if route.get("use_parent_overrides", True):
+            overrides.append(self.app.dependency_overrides)
+        handler = route_app._make_handler(  # pylint: disable=protected-access
             route["func"],
             route["param_names"],
             route["param_converters"],
             route["query_params"],
-            route["expects_body"],
+            route.get("body_param"),
             route["dependencies"],
             route.get("expects_request", False),
             route["method"],
             route["path"],
             route.get("background_param"),
-            [route.get("dependency_overrides", {}), self.app.dependency_overrides],
+            overrides,
         )
         if body is not None and json_body is not None:
             raise ValueError("provide either json_body or body")
         body_bytes = (
             body
             if body is not None
-            else json.dumps(json_body).encode() if json_body else b""
+            else json.dumps(json_body).encode()
+            if json_body
+            else b""
         )
         query = urlencode(params or {}).encode()
-        status, body_str, headers = handler(body_bytes, (), query)
-        return Response(status, body_str, headers)
+        status, body_obj, resp_headers = handler(
+            body_bytes, path_params, query, dict(headers or {})
+        )
+        if isinstance(body_obj, list):
+            text = "".join(body_obj)
+            content = text.encode("latin1")
+            chunks = body_obj
+        elif isinstance(body_obj, bytes):
+            content = body_obj
+            try:
+                text = content.decode()
+            except UnicodeDecodeError:
+                text = content.decode("latin1")
+            chunks = None
+        else:
+            text = body_obj
+            content = text.encode()
+            chunks = None
+        return Response(status, text, resp_headers, content, chunks)
 
-    def get(self, path: str, params: Mapping[str, Any] | None = None) -> Response:
+    def get(
+        self,
+        path: str,
+        params: Mapping[str, Any] | None = None,
+        headers: Mapping[str, str] | None = None,
+    ) -> Response:
         """Send a GET request."""
-        return self.request("GET", path, params=params)
+        return self.request("GET", path, params=params, headers=headers)
 
     def post(
         self,
         path: str,
         json_body: Mapping[str, Any] | None = None,
         params: Mapping[str, Any] | None = None,
+        headers: Mapping[str, str] | None = None,
     ) -> Response:
         """Send a POST request."""
-        return self.request("POST", path, json_body=json_body, params=params)
+        return self.request(
+            "POST", path, json_body=json_body, params=params, headers=headers
+        )
+
+    def head(
+        self,
+        path: str,
+        params: Mapping[str, Any] | None = None,
+        headers: Mapping[str, str] | None = None,
+    ) -> Response:
+        """Send a HEAD request."""
+        return self.request("HEAD", path, params=params, headers=headers)
 
 
 __all__ = ["Response", "TestClient"]
