@@ -2,11 +2,12 @@
 """Test gRPC health check service."""
 
 import pytest
+
+grpc = pytest.importorskip("grpc")
+pytest.importorskip("grpc_health.v1.health_pb2")
 from grpc_health.v1 import health_pb2, health_pb2_grpc
 
 from interfaces.grpc.server import start_grpc_server
-
-grpc = pytest.importorskip("grpc")
 
 
 def test_grpc_health() -> None:
